application/requirements.txt
Flask
Flask-MySQLdb
python-dotenv
mysqlclient
mysql-connector-python
openai
